requirements.txt
Flask
python-dotenv
Werkzeug
Pillow
numpy==2.4.0
opencv-python-headless
huggingface-hub
Pillow
numpy==2.4.0
scikit-image
pymysql
cryptography